api/requirements.txt
asgiref==3.8.1
cachetools==6.0.0
Django==5.2.1
django-rql==4.4.1
djangorestframework==3.16.0
lark-parser==0.11.0
lib-rql==2.0.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
